+++ packages/plugin-page-view-tracking-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.48](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.47...@amplitude/plugin-page-view-tracking-browser@2.3.48) (2025-09-18)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
 ## [2.3.47](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.46...@amplitude/plugin-page-view-tracking-browser@2.3.47) (2025-09-12)
 
 **Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser