+++ packages/segment-session-replay-plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-beta.30](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.29...@amplitude/segment-session-replay-plugin@0.0.0-beta.30) (2025-09-18)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
 # [0.0.0-beta.29](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.28...@amplitude/segment-session-replay-plugin@0.0.0-beta.29) (2025-09-12)
 
 **Note:** Version bump only for package @amplitude/segment-session-replay-plugin